+++ packages/webcomponents/src/api/services/payment.service.ts
@@ -1,22 +1,22 @@
-import { Api, IApiResponseCollection, IApiResponse, IPayment } from '..';
+import { IApiResponseCollection, IApiResponse, IPayment, IPaymentBalanceTransaction } from '..';
+import Api from '../ApiNew';
+
+const api = Api();
 
 export interface IPaymentService {
   fetchPayments(
     accountId: string,
     authToken: string,
-    params: any,
-    apiOrigin?: string
+    params: any
   ): Promise<IApiResponseCollection<IPayment[]>>;
   fetchPayment(
     paymentId: string,
-    authToken: string,
-    apiOrigin?: string
+    authToken: string
   ): Promise<IApiResponse<IPayment>>;
   fetchPaymentTransactions(
     paymentId: string,
-    authToken: string,
     params: any,
-    apiOrigin?: string
+    authToken: string
   ): Promise<IApiResponseCollection<any>>;
 }
 
@@ -24,31 +24,27 @@ export class PaymentService implements IPaymentService {
   async fetchPayments(
     accountId: string,
     authToken: string,
-    params: any,
-    apiOrigin: string = PROXY_API_ORIGIN
+    params: any
   ): Promise<IApiResponseCollection<IPayment[]>> {
-    const api = Api({ authToken, apiOrigin: apiOrigin });
     const endpoint = `account/${accountId}/payments`;
-    return api.get({ endpoint, params });
+    return api.get({ endpoint, params, authToken });
   }
 
   async fetchPayment(
     paymentId: string,
-    authToken: string,
-    apiOrigin: string = PROXY_API_ORIGIN
+    authToken: string
   ): Promise<IApiResponse<IPayment>> {
     const endpoint = `payments/${paymentId}`;
-    return Api({ authToken, apiOrigin }).get({ endpoint });
+    return api.get({ endpoint, authToken });
   }
 
   async fetchPaymentTransactions(
     paymentId: string,
     authToken: string,
-    params: any,
-    apiOrigin: string = PROXY_API_ORIGIN
-  ): Promise<IApiResponseCollection<any>> {
+    params: any
+  ): Promise<IApiResponseCollection<IPaymentBalanceTransaction[]>> {
     const endpoint = `payments/${paymentId}/payment_balance_transactions`;
-    return Api({ authToken, apiOrigin }).get({ endpoint, params });
+    return api.get({ endpoint, params, authToken });
   }
 
 }