+++ packages/webcomponents/src/components/terminal-orders-list/terminal-orders-list-core.tsx
@@ -113,54 +113,52 @@ export class TerminalOrdersListCore {
 
   render() {
     return (
-      <div>
-        <div class="table-wrapper">
-          <table class="table table-hover" part={table}>
-            <thead class="table-head sticky-top">
-              <tr class="table-light text-nowrap">
-                {this.terminalOrdersTable.columnData.map((column) => column)}
+      <div class="table-wrapper">
+        <table class="table table-hover" part={table}>
+          <thead class="table-head sticky-top">
+            <tr class="table-light text-nowrap">
+              {this.terminalOrdersTable.columnData.map((column) => column)}
+            </tr>
+          </thead>
+          <tbody class="table-body">
+            <TableLoadingState
+              columnSpan={this.terminalOrdersTable.columnKeys.length}
+              isLoading={this.loading}
+            />
+            <TableEmptyState
+              isEmpty={this.showEmptyState}
+              columnSpan={this.terminalOrdersTable.columnKeys.length}
+            />
+            <TableErrorState
+              columnSpan={this.terminalOrdersTable.columnKeys.length}
+              errorMessage={this.errorMessage}
+            />
+            {this.showRowData && this.terminalOrdersTable.rowData.map((data, index) => (
+              <tr
+                data-test-id="table-row"
+                data-row-entity-id={this.entityId[index]}
+                onClick={this.rowClickHandler}
+              >
+                {data}
               </tr>
-            </thead>
-            <tbody class="table-body">
-              <TableLoadingState
-                columnSpan={this.terminalOrdersTable.columnKeys.length}
-                isLoading={this.loading}
-              />
-              <TableEmptyState
-                isEmpty={this.showEmptyState}
-                columnSpan={this.terminalOrdersTable.columnKeys.length}
-              />
-              <TableErrorState
-                columnSpan={this.terminalOrdersTable.columnKeys.length}
-                errorMessage={this.errorMessage}
-              />
-              {this.showRowData && this.terminalOrdersTable.rowData.map((data, index) => (
-                <tr
-                  data-test-id="table-row"
-                  data-row-entity-id={this.entityId[index]}
-                  onClick={this.rowClickHandler}
-                >
-                  {data}
-                </tr>
-              ))}
-            </tbody>
-            {this.paging && (
-              <tfoot class="sticky-bottom">
-                <tr class="table-light align-middle">
-                  <td part={tableCell} colSpan={this.terminalOrdersTable.columnData.length}>
-                    <pagination-menu
-                      paging={{
-                        ...this.paging,
-                        handleClickPrevious: this.handleClickPrevious,
-                        handleClickNext: this.handleClickNext,
-                      }}
-                    />
-                  </td>
-                </tr>
-              </tfoot>
-            )}
-          </table>
-        </div>
+            ))}
+          </tbody>
+          {this.paging && (
+            <tfoot class="sticky-bottom">
+              <tr class="table-light align-middle">
+                <td part={tableCell} colSpan={this.terminalOrdersTable.columnData.length}>
+                  <pagination-menu
+                    paging={{
+                      ...this.paging,
+                      handleClickPrevious: this.handleClickPrevious,
+                      handleClickNext: this.handleClickNext,
+                    }}
+                  />
+                </td>
+              </tr>
+            </tfoot>
+          )}
+        </table>
       </div>
     );
   }